2つのMessageActivityを一本化
